ft_trans/backend/game1/requirements.txt
psycopg2-binary>=2.9,<3.0
channels==4.0.0
Django==4.2.6
daphne==4.0.0
PyJWT==2.7.0